refactor: migrate rhog_io, write_elecstat_pot, write_init to source_estate
Move three file pairs from source_io/module_chgpot to source_estate root:
- rhog_io.h/cpp
- write_elecstat_pot.h/cpp
- write_init.h/cpp

Also migrate read_rhog_test.cpp to source_estate/test/.

Update all include paths, CMakeLists.txt (source_io and source_estate,
including test configs), and Makefile.Objects (OBJS_IO -> OBJS_SRCPW).

Add AGENTS.md rules 10-11: forbid #define private public access hacks
in new tests, and require test_<module>.cpp naming for new unit tests.
