+++ source/Makefile.Objects
@@ -614,7 +614,6 @@ OBJS_IO=module_parameter/input_conv.o\
     output.o\
     module_output/print_info.o\
     module_output/read_cube.o\
-    module_chgpot/rhog_io.o\
     module_wf/read_wfc_pw.o\
     module_wf/read_wf2rho_pw.o\
     module_restart/restart.o\
@@ -636,10 +635,8 @@ OBJS_IO=module_parameter/input_conv.o\
     module_output/write_pao.o\
     module_wf/write_wfc_pw.o\
     module_output/write_cube.o\
-    module_chgpot/write_elecstat_pot.o\
     module_elf/write_elf.o\
     module_dipole/write_dipole.o\
-    module_chgpot/write_init.o\
     module_current/td_current_io.o\
     module_current/td_current_io_comm.o\
     td_efield_io.o\
@@ -814,6 +811,9 @@ OBJS_SRCPW=h_ewald_pw.o\
     mix_precond.o\
     charge_mixing_rho.o\
     charge_mixing_uspp.o\
+    rhog_io.o\
+    write_elecstat_pot.o\
+    write_init.o\
     fp_energy.o\
     setup_pot.o\
     setup_pwrho.o\
