+++ source/source_estate/init_scf.cpp
@@ -1,5 +1,5 @@
 #include "elecstate.h"
-#include "source_io/module_chgpot/write_init.h"
+#include "source_estate/write_init.h"
 
 namespace elecstate
 {
